+++ src/components/index.ts
@@ -1,4 +1,5 @@
 export { editableColumnEditorsRegistry } from './ColumnEditor';
+export * from './FieldPicker';
 export * from './Table';
 export * from './TablePanel';
 export * from './TablesEditor';